+++ extensions/volumesnapshot/client/src/main/java/io/fabric8/volumesnapshot/client/internal/VolumeSnapshotOperationsImpl.java
@@ -1,0 +1,49 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.volumesnapshot.client.internal;
+
+import io.fabric8.kubernetes.client.Config;
+import io.fabric8.kubernetes.client.dsl.base.BaseOperation;
+import io.fabric8.kubernetes.client.dsl.base.HasMetadataOperation;
+import io.fabric8.kubernetes.client.dsl.base.OperationContext;
+import io.fabric8.volumesnapshot.api.model.DoneableVolumeSnapshot;
+import io.fabric8.volumesnapshot.api.model.VolumeSnapshot;
+import io.fabric8.volumesnapshot.api.model.VolumeSnapshotList;
+import okhttp3.OkHttpClient;
+
+public class VolumeSnapshotOperationsImpl extends HasMetadataOperation<VolumeSnapshot, VolumeSnapshotList, DoneableVolumeSnapshot, VolumeSnapshotResource> implements VolumeSnapshotResource {
+
+  public VolumeSnapshotOperationsImpl(OkHttpClient client, Config config) {
+    this(new OperationContext().withOkhttpClient(client).withConfig(config));
+  }
+
+  public VolumeSnapshotOperationsImpl(OperationContext ctx) {
+    super(ctx.withApiGroupName("snapshot.storage.k8s.io").withApiGroupVersion("v1beta1").withPlural("volumesnapshots"));
+    this.type = VolumeSnapshot.class;
+    this.listType = VolumeSnapshotList.class;
+    this.doneableType = DoneableVolumeSnapshot.class;
+  }
+
+  @Override
+  public BaseOperation<VolumeSnapshot, VolumeSnapshotList, DoneableVolumeSnapshot, VolumeSnapshotResource> newInstance(OperationContext context) {
+    return new VolumeSnapshotOperationsImpl(context);
+  }
+
+  @Override
+  public boolean isResourceNamespaced() {
+    return true;
+  }
+}