+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix #2292: Update createOrReplace to do replace when create fails with conflict
 
 #### New Features
+* CSI Volume Snapshot extension
 * Fix #2311: Add Support for creating bootstrap project template
 * Fix #2287: Add support for V1 and V1Beta1 CustomResourceDefinition
 * Fix #2319: Create Config without using auto-configure functionality or setting env variables